Updating my changes
Updates to the Join game, no longer a invisible onclick to npc, this time the user needs to go through the window commands. 

If the user already is assigned a player ID, it will let them know that they are already joined. This fixes it so if player2 selects join twice it won't run the OnWelcome function twice or, however, times the users tries. Tricky bugger here.